+++ packages/bunest-adapter/lib/bun.internal.types.ts
@@ -1,9 +1,11 @@
 import {
+  CorsOptions,
   CorsOptionsDelegate,
-  CorsOptions as NestCorsOptions,
 } from '@nestjs/common/interfaces/external/cors-options.interface.js'
 import { Serve, Server, ServerWebSocket, WebSocketHandler } from 'bun'
+import { INestApplication } from '@nestjs/common'
 
+import { BunPreflightHttpServer } from './bun.preflight-http-server.js'
 import { BunRequest } from './bun.request.js'
 
 export interface WsOptions extends Pick<
@@ -16,7 +18,7 @@ export interface WsOptions extends Pick<
   | 'publishToSelf'
   | 'perMessageDeflate'
 > {
-  cors?: true | NestCorsOptions | CorsOptionsDelegate<BunRequest>
+  cors?: true | CorsOptions | CorsOptionsDelegate<BunRequest>
   clientDataFactory?: (req: BunRequest) => unknown
 }
 
@@ -47,4 +49,35 @@ export interface BunWsClientData {
   onCloseInternal?: () => void
   /** Called by NestJS for disconnect handling */
   onDisconnect?: (ws: ServerWebSocket<unknown>) => void
+}
+
+export interface BunStaticAssetsOptions {
+  /**
+   * Enable static assets serving.
+   *
+   * Bun has two distict modes for serving static assets:
+   * 1. Static routes
+   * 2. File routes
+   *
+   * If you set `useStatic: true`, Bun will use static routes for serving assets.
+   * This approach is generally faster for serving static files, as it serves
+   * files directly from memory. However, it comes with some limitations, such as
+   * lack of support for certain features like range requests and directory indexing.
+   * On top of that, static routes didn't respect middlewares due to Bun's internal design.
+   *
+   * On the other hand, if you set `useStatic: false` (the default behavior),
+   * Bun will use file routes, which read files from the filesystem on each request.
+   * This method supports a wider range of features, including range requests, and respects
+   * middlewares. However, it may be slightly slower than static routes due to
+   * filesystem access on each request.
+   *
+   * @see https://bun.com/docs/runtime/http/routing#file-responses-vs-static-responses
+   * @defaults false Use file routes by default.
+   */
+  useStatic?: boolean
+}
+
+export interface NestBunApplication extends INestApplication<BunPreflightHttpServer> {
+  useStaticAssets(path: string, options?: BunStaticAssetsOptions): void
+  enableCors(options?: CorsOptions | CorsOptionsDelegate<BunRequest>): void
 }